requirements.txt

streamlit==1.37.0
pydantic==2.8.2
openai>=1.35.0
pdf2image==1.17.0
pytesseract==0.3.10
pdfplumber==0.11.4
opencv-python==4.10.0.84
python-dotenv==1.0.1
rapidfuzz==3.9.6
tenacity==8.2.3
tqdm==4.66.5
pytest==8.3.2
